+++ lib/iban.rb
@@ -1,6 +1,0 @@
-require 'iban_bru/version'
-require 'iban_bru/iban'
-require 'iban_bru/iban_builder'
-
-module IbanBru
-end