+++ lib/ibandit/iban_builder.rb
@@ -1,4 +1,4 @@
-module IbanBru
+module Ibandit
   module IBANBuilder
     SUPPORTED_COUNTRY_CODES = %w(AT ES IT FR PT MC SM BE EE CY FI LU LV SI SK)