+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## 0.0.2 - December 4, 2014
 
-- Rename to iban-bru
+- Rename to ibandit
 
 ## 0.0.1 - November 27, 2014